+++ master/docs/release-notes.rst
@@ -14,6 +14,9 @@ Master
 Deprecations, Removals, and Non-Compatible Changes
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
+* ``buildbot start`` no longer invokes make if a ``Makefile.buildbot`` exists.
+  If you are using this functionality, consider invoking make directly.
+
 Features
 ~~~~~~~~